create script tooling to create all redirects
for redirect to work, all the pages you want to redirect will need objects in same folder we server from

assume docs_preview is the 'final' place we will serve from - all the redirect files will also have to be created in that folder to work

do first 10 redirects for testing. remove the --start --end to do all.

we wont want to run it all the time - so comment it out once all redirects are in place. we can figure a more robust check if this becomes an issue.

llm: o3